segregate print_query_response_data
