+++ bin/src/ctl/command.rs
@@ -16,7 +16,7 @@ use crate::{
         create_channel,
         display::{
             create_queried_application_table, print_frontend_list, print_json_response,
-            print_query_answers,
+            print_query_answers, print_query_response_data,
         },
         CommandManager,
     },
@@ -718,246 +718,7 @@ impl CommandManager {
                 bail!("could not query proxy state: {}", message.message);
             }
             CommandStatus::Ok => {
-                if let Some(needle) = application_id.or(domain) {
-                    if let Some(CommandResponseData::Query(data)) = message.data {
-                        if json {
-                            return Ok(print_json_response(&data)?);
-                        }
-
-                        let application_headers = vec!["id", "sticky_session", "https_redirect"];
-                        let mut application_table =
-                            create_queried_application_table(application_headers, &data);
-
-                        let http_headers = vec!["id", "hostname", "path"];
-                        let mut frontend_table =
-                            create_queried_application_table(http_headers, &data);
-
-                        let https_headers = vec!["id", "hostname", "path"];
-                        let mut https_frontend_table =
-                            create_queried_application_table(https_headers, &data);
-
-                        let tcp_headers = vec!["id", "address"];
-                        let mut tcp_frontend_table =
-                            create_queried_application_table(tcp_headers, &data);
-
-                        let backend_headers = vec!["backend id", "IP address", "Backup"];
-                        let mut backend_table =
-                            create_queried_application_table(backend_headers, &data);
-
-                        let keys: HashSet<&String> = data.keys().collect();
-
-                        let mut application_data = HashMap::new();
-                        let mut frontend_data = HashMap::new();
-                        let mut https_frontend_data = HashMap::new();
-                        let mut tcp_frontend_data = HashMap::new();
-                        let mut backend_data = HashMap::new();
-
-                        for (ref key, ref metrics) in data.iter() {
-                            //let m: u8 = metrics;
-                            if let &QueryAnswer::Applications(ref apps) = *metrics {
-                                for app in apps.iter() {
-                                    let entry = application_data.entry(app).or_insert(Vec::new());
-                                    entry.push((*key).clone());
-
-                                    for frontend in app.http_frontends.iter() {
-                                        let entry =
-                                            frontend_data.entry(frontend).or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-
-                                    for frontend in app.https_frontends.iter() {
-                                        let entry = https_frontend_data
-                                            .entry(frontend)
-                                            .or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-
-                                    for frontend in app.tcp_frontends.iter() {
-                                        let entry =
-                                            tcp_frontend_data.entry(frontend).or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-
-                                    for backend in app.backends.iter() {
-                                        let entry =
-                                            backend_data.entry(backend).or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-                                }
-                            }
-                        }
-
-                        println!("Cluster level configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in application_data.iter() {
-                            let mut row = Vec::new();
-                            row.push(cell!(key
-                                .configuration
-                                .clone()
-                                .map(|conf| conf.cluster_id)
-                                .unwrap_or(String::from(""))));
-                            row.push(cell!(key
-                                .configuration
-                                .clone()
-                                .map(|conf| conf.sticky_session)
-                                .unwrap_or(false)));
-                            row.push(cell!(key
-                                .configuration
-                                .clone()
-                                .map(|conf| conf.https_redirect)
-                                .unwrap_or(false)));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            application_table.add_row(Row::new(row));
-                        }
-
-                        application_table.printstd();
-
-                        println!("\nHTTP frontends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in frontend_data.iter() {
-                            let mut row = Vec::new();
-                            match &key.route {
-                                Route::ClusterId(cluster_id) => row.push(cell!(cluster_id)),
-                                Route::Deny => row.push(cell!("-")),
-                            }
-                            row.push(cell!(key.hostname));
-                            row.push(cell!(key.path));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            frontend_table.add_row(Row::new(row));
-                        }
-
-                        frontend_table.printstd();
-
-                        println!("\nHTTPS frontends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in https_frontend_data.iter() {
-                            let mut row = Vec::new();
-                            match &key.route {
-                                Route::ClusterId(cluster_id) => row.push(cell!(cluster_id)),
-                                Route::Deny => row.push(cell!("-")),
-                            }
-                            row.push(cell!(key.hostname));
-                            row.push(cell!(key.path));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            https_frontend_table.add_row(Row::new(row));
-                        }
-
-                        https_frontend_table.printstd();
-
-                        println!("\nTCP frontends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in tcp_frontend_data.iter() {
-                            let mut row = Vec::new();
-                            row.push(cell!(key.cluster_id));
-                            row.push(cell!(format!("{}", key.address)));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            tcp_frontend_table.add_row(Row::new(row));
-                        }
-
-                        tcp_frontend_table.printstd();
-
-                        println!("\nbackends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in backend_data.iter() {
-                            let mut row = Vec::new();
-                            let backend_backup =
-                                key.backup.map(|b| if b { "X" } else { "" }).unwrap_or("");
-                            row.push(cell!(key.backend_id));
-                            row.push(cell!(format!("{}", key.address)));
-                            row.push(cell!(backend_backup));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            backend_table.add_row(Row::new(row));
-                        }
-
-                        backend_table.printstd();
-                    }
-                } else {
-                    if let Some(CommandResponseData::Query(data)) = message.data {
-                        let mut table = Table::new();
-                        let mut header = Vec::new();
-                        header.push(cell!("key"));
-                        for ref key in data.keys() {
-                            header.push(cell!(&key));
-                        }
-                        header.push(cell!("desynchronized"));
-                        table.add_row(Row::new(header));
-
-                        let mut query_data = HashMap::new();
-
-                        for ref metrics in data.values() {
-                            //let m: u8 = metrics;
-                            if let &QueryAnswer::ApplicationsHashes(ref apps) = *metrics {
-                                for (ref key, ref value) in apps.iter() {
-                                    (*(query_data.entry((*key).clone()).or_insert(Vec::new())))
-                                        .push(*value);
-                                }
-                            }
-                        }
-
-                        for (ref key, ref values) in query_data.iter() {
-                            let mut row = Vec::new();
-                            row.push(cell!(key));
-
-                            for val in values.iter() {
-                                row.push(cell!(format!("{}", val)));
-                            }
-
-                            let hs: HashSet<&u64> = values.iter().cloned().collect();
-
-                            let diff = hs.len() > 1;
-
-                            if diff {
-                                row.push(cell!(String::from("X")));
-                            } else {
-                                row.push(cell!(String::from("")));
-                            }
-
-                            table.add_row(Row::new(row));
-                        }
-
-                        table.printstd();
-                    }
-                }
+                print_query_response_data(application_id, domain, message.data, json)?;
             }
         }
